Optionally mount scratch directory for feature install RUN
We don't need to bind-mount the scratch directory when running a build
with envbuilder, since its in-container build has access to the whole
root filesystem of the container where envbuilder runs. However, if
external code calls into the `devcontainer` package to produce a
Dockerfile to use for a non-kaniko-based build, the resulting Dockerfile
needs to bind-mount the scratch directory in order to access the feature
install script and associated files.